requirements.txt
gymnasium==1.2.0
hydra-core==1.3.2
matplotlib==3.10.5
omegaconf==2.3.0
smac==2.3.1
tqdm==4.67.1
